# riski-oparl/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RISKI OParl API",
    "summary": "OParl 1.1 publishing API for Munich's political information system (RIS).",
    "description": "\nThis service publishes the RISKI database according to the\n[OParl 1.1 specification](https://oparl.org).\n\n### How to use it\nOParl is **link-driven** — start at the entry point and follow the URLs:\n\n1. `GET /oparl/v1/system` — the entry point. Its `body` field links to the body list.\n2. `GET /oparl/v1/bodies` — pick a body; its `organization` / `person` / `meeting` /\n   `paper` fields are URLs to paginated **external lists**.\n3. Follow object URLs (e.g. a paper's `originatorPerson`) to fetch individual objects.\n\nEvery object's `id` is the canonical URL on this API; the original RIS web page is\nprovided in the `web` field.\n\n### External lists\nSupport pagination via `page` and time filtering via `created_since`, `created_until`,\n`modified_since`, `modified_until`. Responses contain `data`, `pagination` and `links`\n(`first` / `prev` / `next` / `last`).\n",
    "contact": {
      "name": "Landeshauptstadt München",
      "url": "https://github.com/it-at-m/riski"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/oparl/v1/bodies": {
      "get": {
        "tags": [
          "OParl: Body"
        ],
        "summary": "List bodies",
        "description": "Paginated external list of `Body` objects.",
        "operationId": "list_bodies_oparl_v1_bodies_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "created_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects created at/after this ISO-8601 timestamp.",
              "examples": [
                "2026-01-01T00:00:00"
              ],
              "title": "Created Since"
            },
            "description": "Only objects created at/after this ISO-8601 timestamp."
          },
          {
            "name": "created_until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects created at/before this ISO-8601 timestamp.",
              "title": "Created Until"
            },
            "description": "Only objects created at/before this ISO-8601 timestamp."
          },
          {
            "name": "modified_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects modified at/after this ISO-8601 timestamp (incremental harvesting).",
              "examples": [
                "2026-06-01T00:00:00"
              ],
              "title": "Modified Since"
            },
            "description": "Only objects modified at/after this ISO-8601 timestamp (incremental harvesting)."
          },
          {
            "name": "modified_until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects modified at/before this ISO-8601 timestamp.",
              "title": "Modified Until"
            },
            "description": "Only objects modified at/before this ISO-8601 timestamp."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "data": [
                    {
                      "id": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261",
                      "type": "https://schema.oparl.org/1.1/Body",
                      "system": "https://oparl.example.org/oparl/v1/system",
                      "name": "Landeshauptstadt München",
                      "shortName": "München",
                      "website": "https://risi.muenchen.de/risi",
                      "license": "https://www.govdata.de/dl-de/by-2-0",
                      "organization": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/organizations",
                      "person": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/people",
                      "meeting": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/meetings",
                      "paper": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/papers",
                      "agendaItem": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/agendaItems",
                      "consultation": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/consultations",
                      "file": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/files",
                      "locationList": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/locations",
                      "membership": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/memberships",
                      "legislativeTerm": [],
                      "created": "2026-06-24T15:47:08+02:00",
                      "modified": "2026-06-24T15:47:08+02:00"
                    }
                  ],
                  "pagination": {
                    "totalElements": 1,
                    "elementsPerPage": 50,
                    "currentPage": 1,
                    "totalPages": 1
                  },
                  "links": {
                    "first": "https://oparl.example.org/oparl/v1/list?page=1",
                    "last": "https://oparl.example.org/oparl/v1/list?page=1"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oparl/v1/bodies/{db_id}": {
      "get": {
        "tags": [
          "OParl: Body"
        ],
        "summary": "Get a body",
        "operationId": "get_body_oparl_v1_bodies__db_id__get",
        "parameters": [
          {
            "name": "db_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Db Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "id": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261",
                  "type": "https://schema.oparl.org/1.1/Body",
                  "system": "https://oparl.example.org/oparl/v1/system",
                  "name": "Landeshauptstadt München",
                  "shortName": "München",
                  "website": "https://risi.muenchen.de/risi",
                  "license": "https://www.govdata.de/dl-de/by-2-0",
                  "organization": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/organizations",
                  "person": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/people",
                  "meeting": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/meetings",
                  "paper": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/papers",
                  "agendaItem": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/agendaItems",
                  "consultation": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/consultations",
                  "file": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/files",
                  "locationList": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/locations",
                  "membership": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261/memberships",
                  "legislativeTerm": [],
                  "created": "2026-06-24T15:47:08+02:00",
                  "modified": "2026-06-24T15:47:08+02:00"
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "type": "https://schema.oparl.org/1.1/Error",
                    "name": "Not Found",
                    "message": "No Paper with id 00000000-0000-0000-0000-000000000000."
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oparl/v1/bodies/{db_id}/organizations": {
      "get": {
        "tags": [
          "OParl: Lists"
        ],
        "summary": "List organizations of a body",
        "operationId": "list_body_organizations_oparl_v1_bodies__db_id__organizations_get",
        "parameters": [
          {
            "name": "db_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Db Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "created_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects created at/after this ISO-8601 timestamp.",
              "examples": [
                "2026-01-01T00:00:00"
              ],
              "title": "Created Since"
            },
            "description": "Only objects created at/after this ISO-8601 timestamp."
          },
          {
            "name": "created_until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects created at/before this ISO-8601 timestamp.",
              "title": "Created Until"
            },
            "description": "Only objects created at/before this ISO-8601 timestamp."
          },
          {
            "name": "modified_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects modified at/after this ISO-8601 timestamp (incremental harvesting).",
              "examples": [
                "2026-06-01T00:00:00"
              ],
              "title": "Modified Since"
            },
            "description": "Only objects modified at/after this ISO-8601 timestamp (incremental harvesting)."
          },
          {
            "name": "modified_until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only objects modified at/before this ISO-8601 timestamp.",
              "title": "Modified Until"
            },
            "description": "Only objects modified at/before this ISO-8601 timestamp."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "data": [
                    {
                      "id": "https://oparl.example.org/oparl/v1/organizations/3a1b8e10-0000-4000-8000-000000000001",
                      "type": "https://schema.oparl.org/1.1/Organization",
                      "body": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261",
                      "name": "Fraktion Beispiel",
                      "shortName": "FB",
                      "organizationType": "Fraktion",
                      "classification": "Fraktion",
                      "membership": [
                        "https://oparl.example.org/oparl/v1/memberships/9c1f0000-0000-4000-8000-000000000001"
                      ]
                    }
                  ],
                  "pagination": {
                    "totalElements": 19,
                    "elementsPerPage": 50,
                    "currentPage": 1,
                    "totalPages": 1
                  },
                  "links": {
                    "first": "https://oparl.example.org/oparl/v1/list?page=1",
                    "last": "https://oparl.example.org/oparl/v1/list?page=1"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/oparl/v1/organizations/{db_id}": {
      "get": {
        "tags": [
          "OParl: Objects"
        ],
        "summary": "Get an organization",
        "operationId": "get_organization_oparl_v1_organizations__db_id__get",
        "parameters": [
          {
            "name": "db_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Db Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "id": "https://oparl.example.org/oparl/v1/organizations/3a1b8e10-0000-4000-8000-000000000001",
                  "type": "https://schema.oparl.org/1.1/Organization",
                  "body": "https://oparl.example.org/oparl/v1/bodies/dbf2fbcb-38b4-4334-93fa-875f9963b261",
                  "name": "Fraktion Beispiel",
                  "shortName": "FB",
                  "organizationType": "Fraktion",
                  "classification": "Fraktion",
                  "membership": [
                    "https://oparl.example.org/oparl/v1/memberships/9c1f0000-0000-4000-8000-000000000001"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "type": "https://schema.oparl.org/1.1/Error",
                    "name": "Not Found",
                    "message": "No Paper with id 00000000-0000-0000-0000-000000000000."
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}